+++ fmi-schema/src/fmi2/attribute_groups.rs
@@ -8,7 +8,8 @@ pub struct RealAttributes {
     #[yaserde(attribute)]
     pub unit: Option<String>,
 
-    /// Default display unit, provided the conversion of values in "unit" to values in "displayUnit" is defined in UnitDefinitions / Unit / DisplayUnit.
+    /// Default display unit, provided the conversion of values in "unit" to values in
+    /// "displayUnit" is defined in UnitDefinitions / Unit / DisplayUnit.
     #[yaserde(attribute, rename = "displayUnit")]
     pub display_unit: Option<String>,
 
@@ -27,7 +28,8 @@ pub struct RealAttributes {
     #[yaserde(attribute, rename = "nominal")]
     pub nominal: Option<f64>,
 
-    /// Set to true, e.g., for crank angle. If true and variable is a state, relative tolerance should be zero on this variable.
+    /// Set to true, e.g., for crank angle. If true and variable is a state, relative tolerance
+    /// should be zero on this variable.
     #[yaserde(attribute, rename = "unbounded")]
     pub unbounded: bool,
 }
