+++ fmi-schema/src/fmi2/variable_dependency.rs
@@ -1,7 +1,7 @@
 use yaserde_derive::{YaDeserialize, YaSerialize};
 
-/// Dependency of scalar Unknown from Knowns in Continuous-Time and Event Mode (ModelExchange), and at Communication
-/// Points (CoSimulation): Unknown=f(Known_1, Known_2, ...).
+/// Dependency of scalar Unknown from Knowns in Continuous-Time and Event Mode (ModelExchange), and
+/// at Communication Points (CoSimulation): Unknown=f(Known_1, Known_2, ...).
 /// The Knowns are "inputs", "continuous states" and "independent variable" (usually time)".
 #[derive(Default, PartialEq, Debug, YaSerialize, YaDeserialize)]
 pub struct Fmi2VariableDependency {
@@ -9,17 +9,19 @@ pub struct Fmi2VariableDependency {
     #[yaserde(attribute)]
     pub index: u32,
 
-    /// Defines the dependency of the Unknown (directly or indirectly via auxiliary variables) on the Knowns in
-    /// Continuous-Time and Event Mode ([`super::ModelExchange`]) and at Communication Points ([`super::CoSimulation`]).
+    /// Defines the dependency of the Unknown (directly or indirectly via auxiliary variables) on
+    /// the Knowns in Continuous-Time and Event Mode ([`super::ModelExchange`]) and at
+    /// Communication Points ([`super::CoSimulation`]).
     ///
-    /// If not present, it must be assumed that the Unknown depends on all Knowns. If present as empty list, the Unknown
-    /// depends on none of the Knowns. Otherwise the Unknown depends on the Knowns defined by the given
-    /// [`super::ScalarVariable`] indices. The indices are ordered according to size, starting with the smallest index.
+    /// If not present, it must be assumed that the Unknown depends on all Knowns. If present as
+    /// empty list, the Unknown depends on none of the Knowns. Otherwise the Unknown depends on
+    /// the Knowns defined by the given [`super::ScalarVariable`] indices. The indices are
+    /// ordered according to size, starting with the smallest index.
     #[yaserde(attribute, rename = "dependencies")]
     pub dependencies: Vec<u32>,
 
-    /// If not present, it must be assumed that the Unknown depends on the Knowns without a particular structure.
-    /// Otherwise, the corresponding Known v enters the equation as:
+    /// If not present, it must be assumed that the Unknown depends on the Knowns without a
+    /// particular structure. Otherwise, the corresponding Known v enters the equation as:
     ///
     /// * [`DependenciesKind::Dependent`]:  no particular structure, f(v)
     /// * [`DependenciesKind::Constant`]:   constant factor, c*v (only for Real variablse)
@@ -27,7 +29,8 @@ pub struct Fmi2VariableDependency {
     /// * [`DependenciesKind::Tunable`]:    tunable factor, p*v (only for Real variables)
     /// * [`DependenciesKind::Discrete`]:   discrete factor, d*v (only for Real variables)
     ///
-    /// If [`Self::dependencies_kind`] is present, [`Self::dependencies`] must be present and must have the same number of list elements.
+    /// If [`Self::dependencies_kind`] is present, [`Self::dependencies`] must be present and must
+    /// have the same number of list elements.
     #[yaserde(child, attribute, rename = "dependenciesKind")]
     pub dependencies_kind: Vec<DependenciesKind>,
 }
