+++ fmi-schema/src/lib.rs
@@ -1,5 +1,4 @@
 #![doc=include_str!( "../README.md")]
-//!
 //! ## Feature flags
 #![doc = document_features::document_features!()]
 #![deny(unsafe_code)]
