+++ rustfmt.toml
@@ -1,3 +1,3 @@
 wrap_comments = true
-comment_width = 180
+comment_width = 120
 normalize_comments = true
