+++ src/fmi2/instance/mod.rs
@@ -1,12 +1,11 @@
 //! FMI 2.0 instance interface
 
-use crate::fmi2::instance::traits::Common;
+use crate::{fmi2::instance::traits::Common, Error};
 
 use super::{binding, logger, schema, CallbackFunctions, Fmi2Status};
 
-use std::ffi::CString;
-
 mod co_simulation;
+mod common;
 mod model_exchange;
 pub mod traits;
 
@@ -49,40 +48,36 @@ impl Default for CallbackFunctions {
     }
 }
 
-/// Check the internal consistency of the FMU by comparing the TypesPlatform and FMI versions
-/// from the library and the Model Description XML
-#[cfg(feature = "disable")]
-fn check_consistency(import: &Import, common: &FmiCommon) -> FmiResult<()> {
-    let types_platform =
-        unsafe { std::ffi::CStr::from_ptr(common.get_types_platform()) }.to_bytes_with_nul();
+impl<'a, Tag> Instance<'a, Tag> {
+    /// Check the internal consistency of the FMU by comparing the TypesPlatform and FMI versions
+    /// from the library and the Model Description XML
+    pub fn check_consistency(&self) -> Result<(), Error> {
+        //TODO: Fix
+        //let types_platform = self.get_types_platform();
+        //if types_platform != binding::fmi2TypesPlatform {
+        //    return Err(FmiError::TypesPlatformMismatch {
+        //        found: types_platform.into(),
+        //    });
+        //}
 
-    if types_platform != binding::fmi2TypesPlatform {
-        return Err(FmiError::TypesPlatformMismatch {
-            found: types_platform.into(),
-        });
+        let fmi_version = self.version();
+        if fmi_version != self.schema.fmi_version {
+            return Err(Error::FmiVersionMismatch {
+                found: fmi_version.to_owned(),
+                expected: self.schema.fmi_version.to_owned(),
+            });
+        }
+
+        Ok(())
     }
-
-    let fmi_version = unsafe { std::ffi::CStr::from_ptr(common.get_version()) }.to_bytes();
-    if fmi_version != import.descr().fmi_version.as_bytes() {
-        return Err(FmiError::FmiVersionMismatch {
-            found: fmi_version.into(),
-            expected: import.descr().fmi_version.as_bytes().into(),
-        });
-    }
-
-    Ok(())
 }
-
-// We assume here that the exported FMUs are thread-safe (true for OpenModelica)
-//unsafe impl<A: FmiApi> Send for Instance<A> {}
-//unsafe impl<A: FmiApi> Sync for Instance<A> {}
 
 /// FmuState wraps the FMUstate pointer and is used for managing FMU state
 #[cfg(feature = "disable")]
 pub struct FmuState<'a, Tag> {
     component: &'a Instance<'a, Tag>,
     state: binding::fmi2FMUstate,
-    //container: &'a dlopen::wrapper::Container<A>,
+    // container: &'a dlopen::wrapper::Container<A>,
 }
 
 #[cfg(feature = "disable")]
@@ -98,193 +93,6 @@ impl<'a, Tag> Drop for FmuState<'a, Tag> {
                 .fmi2FreeFMUstate(self.component.component, self.state)
         };
         todo!();
-    }
-}
-
-impl<'a, A> traits::Common for Instance<'a, A> {
-    fn name(&self) -> &str {
-        &self.schema.model_name
-    }
-
-    fn version(&self) -> &str {
-        unsafe { std::ffi::CStr::from_ptr(self.binding.fmi2GetVersion()) }
-            .to_str()
-            .expect("Error converting string")
-    }
-
-    fn set_debug_logging(&self, logging_on: bool, categories: &[&str]) -> Fmi2Status {
-        let category_cstr = categories
-            .iter()
-            .map(|c| CString::new(*c).unwrap())
-            .collect::<Vec<_>>();
-
-        let category_ptrs: Vec<_> = category_cstr.iter().map(|c| c.as_ptr()).collect();
-
-        Fmi2Status(unsafe {
-            self.binding.fmi2SetDebugLogging(
-                self.component,
-                logging_on as binding::fmi2Boolean,
-                category_ptrs.len(),
-                category_ptrs.as_ptr(),
-            )
-        })
-    }
-
-    fn setup_experiment(
-        &self,
-        tolerance: Option<f64>,
-        start_time: f64,
-        stop_time: Option<f64>,
-    ) -> Fmi2Status {
-        Fmi2Status(unsafe {
-            self.binding.fmi2SetupExperiment(
-                self.component,
-                tolerance.is_some() as binding::fmi2Boolean,
-                tolerance.unwrap_or(0.0),
-                start_time,
-                stop_time.is_some() as binding::fmi2Boolean,
-                stop_time.unwrap_or(0.0),
-            )
-        })
-    }
-
-    fn enter_initialization_mode(&self) -> Fmi2Status {
-        Fmi2Status(unsafe { self.binding.fmi2EnterInitializationMode(self.component) })
-    }
-
-    fn exit_initialization_mode(&self) -> Fmi2Status {
-        Fmi2Status(unsafe { self.binding.fmi2ExitInitializationMode(self.component) })
-    }
-
-    fn terminate(&self) -> Fmi2Status {
-        Fmi2Status(unsafe { self.binding.fmi2Terminate(self.component) })
-    }
-
-    fn reset(&self) -> Fmi2Status {
-        Fmi2Status(unsafe { self.binding.fmi2Reset(self.component) })
-    }
-
-    fn get_real(
-        &self,
-        vrs: &[binding::fmi2ValueReference],
-        values: &mut [binding::fmi2Real],
-    ) -> Fmi2Status {
-        assert_eq!(vrs.len(), values.len());
-        Fmi2Status(unsafe {
-            self.binding
-                .fmi2GetReal(self.component, vrs.as_ptr(), vrs.len(), values.as_mut_ptr())
-        })
-    }
-
-    fn get_integer(
-        &self,
-        vrs: &[binding::fmi2ValueReference],
-        values: &mut [binding::fmi2Integer],
-    ) -> Fmi2Status {
-        Fmi2Status(unsafe {
-            self.binding.fmi2GetInteger(
-                self.component,
-                vrs.as_ptr(),
-                vrs.len(),
-                values.as_mut_ptr(),
-            )
-        })
-    }
-
-    fn get_boolean(
-        &self,
-        sv: &[binding::fmi2ValueReference],
-        v: &mut [binding::fmi2Boolean],
-    ) -> Fmi2Status {
-        Fmi2Status(unsafe {
-            self.binding
-                .fmi2GetBoolean(self.component, sv.as_ptr(), sv.len(), v.as_mut_ptr())
-        })
-    }
-
-    fn get_string(
-        &self,
-        sv: &[binding::fmi2ValueReference],
-        v: &mut [binding::fmi2String],
-    ) -> Fmi2Status {
-        Fmi2Status(unsafe {
-            self.binding
-                .fmi2GetString(self.component, sv.as_ptr(), sv.len(), v.as_mut_ptr())
-        })
-    }
-
-    fn set_real(
-        &self,
-        vrs: &[binding::fmi2ValueReference],
-        values: &[binding::fmi2Real],
-    ) -> Fmi2Status {
-        assert_eq!(vrs.len(), values.len());
-        Fmi2Status(unsafe {
-            self.binding.fmi2SetReal(
-                self.component,
-                vrs.as_ptr() as *const u32,
-                values.len(),
-                values.as_ptr(),
-            )
-        })
-    }
-
-    fn set_integer(
-        &self,
-        vrs: &[binding::fmi2ValueReference],
-        values: &[binding::fmi2Integer],
-    ) -> Fmi2Status {
-        assert_eq!(vrs.len(), values.len());
-        Fmi2Status(unsafe {
-            self.binding
-                .fmi2SetInteger(self.component, vrs.as_ptr(), values.len(), values.as_ptr())
-        })
-    }
-
-    fn set_boolean(
-        &self,
-        vrs: &[binding::fmi2ValueReference],
-        values: &mut [binding::fmi2Boolean],
-    ) -> Fmi2Status {
-        assert_eq!(vrs.len(), values.len());
-        Fmi2Status(unsafe {
-            self.binding
-                .fmi2SetBoolean(self.component, vrs.as_ptr(), values.len(), values.as_ptr())
-        })
-    }
-
-    fn set_string(
-        &self,
-        _vrs: &[binding::fmi2ValueReference],
-        _values: &[binding::fmi2String],
-    ) -> Fmi2Status {
-        unimplemented!()
-    }
-
-    // fn get_fmu_state(&self, state: *mut fmi2FMUstate) -> FmiResult<()> {}
-
-    // fn set_fmu_state(&self, state: &[u8]) -> FmiResult<()> {}
-
-    fn get_directional_derivative(
-        &self,
-        unknown_vrs: &[binding::fmi2ValueReference],
-        known_vrs: &[binding::fmi2ValueReference],
-        dv_known_values: &[binding::fmi2Real],
-        dv_unknown_values: &mut [binding::fmi2Real],
-    ) -> Fmi2Status {
-        assert!(unknown_vrs.len() == dv_unknown_values.len());
-        assert!(known_vrs.len() == dv_unknown_values.len());
-        Fmi2Status(unsafe {
-            self.binding.fmi2GetDirectionalDerivative(
-                self.component,
-                unknown_vrs.as_ptr(),
-                unknown_vrs.len(),
-                known_vrs.as_ptr(),
-                known_vrs.len(),
-                dv_known_values.as_ptr(),
-                dv_unknown_values.as_mut_ptr(),
-            )
-        })
     }
 }
