+++ src/fmi2/mod.rs
@@ -56,37 +56,45 @@ pub enum StatusKind {
 pub enum Fmi2Res {
     /// All well
     OK,
-    /// Things are not quite right, but the computation can continue. Function “logger” was called in the model (see
-    /// below), and it is expected that this function has shown the prepared information message to the user.
+    /// Things are not quite right, but the computation can continue. Function “logger” was called
+    /// in the model (see below), and it is expected that this function has shown the prepared
+    /// information message to the user.
     Warning,
-    /// This status is returned only from the co-simulation interface, if the slave executes the function in an
-    /// asynchronous way. That means the slave starts to compute but returns immediately.
+    /// This status is returned only from the co-simulation interface, if the slave executes the
+    /// function in an asynchronous way. That means the slave starts to compute but returns
+    /// immediately.
     ///
-    /// The master has to call [`instance::traits::CoSimulation::get_status`](..., fmi2DoStepStatus) to determine if
-    /// the slave has finished the computation. Can be returned only by [`instance::traits::CoSimulation::do_step`]
+    /// The master has to call [`instance::traits::CoSimulation::get_status`](...,
+    /// fmi2DoStepStatus) to determine if the slave has finished the computation. Can be
+    /// returned only by [`instance::traits::CoSimulation::do_step`]
     /// and by [`instance::traits::CoSimulation::get_status`].
     Pending,
 }
 
 #[derive(Debug, thiserror::Error)]
-pub enum Fmi2Err {
-    /// For “model exchange”: It is recommended to perform a smaller step size and evaluate the model equations again,
-    /// for example because an iterative solver in the model did not converge or because a function is outside of its
-    /// domain (for example sqrt(<negative number>)). If this is not possible, the simulation has to be terminated.
+pub enum Fmi2Error {
+    /// For “model exchange”: It is recommended to perform a smaller step size and evaluate the
+    /// model equations again, for example because an iterative solver in the model did not
+    /// converge or because a function is outside of its domain (for example sqrt(<negative
+    /// number>)). If this is not possible, the simulation has to be terminated.
     ///
-    /// For “co-simulation”: [`Fmi2Err::Discard`] is returned also if the slave is not able to return the required
-    /// status information. The master has to decide if the simulation run can be continued.
+    /// For “co-simulation”: [`Fmi2Err::Discard`] is returned also if the slave is not able to
+    /// return the required status information. The master has to decide if the simulation run
+    /// can be continued.
     ///
-    /// In both cases, function “logger” was called in the FMU (see below) and it is expected that this function has
-    /// shown the prepared information message to the user if the FMU was called in debug mode (loggingOn = true).
-    /// Otherwise, “logger” should not show a message.
+    /// In both cases, function “logger” was called in the FMU (see below) and it is expected that
+    /// this function has shown the prepared information message to the user if the FMU was
+    /// called in debug mode (loggingOn = true). Otherwise, “logger” should not show a message.
     #[error("Discard")]
     Discard,
-    /// The FMU encountered an error. The simulation cannot be continued with this FMU instance. If one of the functions
-    /// returns [`Fmi2Err::Error`], it can be tried to restart the simulation from a formerly stored FMU state by
-    /// calling [`instance::traits::Common::set_fmu_state`]. This can be done if the capability flag `can_get_and_set_fmu_state`
-    /// is true and [`instance::traits::Common::get_fmu_state`] was called before in non-erroneous state. If not, the simulation
-    /// cannot be continued and [`instance::traits::Common::reset`] must be called afterwards.
+    /// The FMU encountered an error. The simulation cannot be continued with this FMU instance. If
+    /// one of the functions returns [`Fmi2Err::Error`], it can be tried to restart the
+    /// simulation from a formerly stored FMU state by
+    /// calling [`instance::traits::Common::set_fmu_state`]. This can be done if the capability
+    /// flag `can_get_and_set_fmu_state` is true and
+    /// [`instance::traits::Common::get_fmu_state`] was called before in non-erroneous state. If
+    /// not, the simulation cannot be continued and [`instance::traits::Common::reset`] must be
+    /// called afterwards.
     #[error("Error")]
     Error,
     /// The model computations are irreparably corrupted for all FMU instances.
@@ -100,7 +108,7 @@ pub struct Fmi2Status(binding::fmi2Status);
 impl Fmi2Status {
     /// Convert to [`Result<Fmi2Res, Fmi2Err>`]
     #[inline]
-    pub fn ok(self) -> Result<Fmi2Res, Fmi2Err> {
+    pub fn ok(self) -> Result<Fmi2Res, Fmi2Error> {
         self.into()
     }
 }
@@ -111,14 +119,14 @@ impl From<binding::fmi2Status> for Fmi2Status {
     }
 }
 
-impl From<Fmi2Status> for Result<Fmi2Res, Fmi2Err> {
+impl From<Fmi2Status> for Result<Fmi2Res, Fmi2Error> {
     fn from(Fmi2Status(status): Fmi2Status) -> Self {
         match status {
             binding::fmi2Status_fmi2OK => Ok(Fmi2Res::OK),
             binding::fmi2Status_fmi2Warning => Ok(Fmi2Res::Warning),
-            binding::fmi2Status_fmi2Discard => Err(Fmi2Err::Discard),
-            binding::fmi2Status_fmi2Error => Err(Fmi2Err::Error),
-            binding::fmi2Status_fmi2Fatal => Err(Fmi2Err::Fatal),
+            binding::fmi2Status_fmi2Discard => Err(Fmi2Error::Discard),
+            binding::fmi2Status_fmi2Error => Err(Fmi2Error::Error),
+            binding::fmi2Status_fmi2Fatal => Err(Fmi2Error::Fatal),
             _ => unreachable!("Invalid status"),
         }
     }
