+++ src/fmi3/instance/mod.rs
@@ -22,7 +22,7 @@ pub struct Instance<'a, Tag> {
     instance: binding::fmi3Instance,
     /// Derived model description
     model: &'a schema::FmiModelDescription,
-    //model: &'a model::ModelDescription,
+    // model: &'a model::ModelDescription,
     _tag: std::marker::PhantomData<&'a Tag>,
 }
 
@@ -55,18 +55,21 @@ impl<'a, Tag> Drop for Fmu3State<'a, Tag> {
 /// Return value of [`Common::update_discrete_states()`]
 #[derive(Default, Debug, PartialEq)]
 pub struct DiscreteStates {
-    /// The importer must stay in Event Mode for another event iteration, starting a new super-dense time instant.
+    /// The importer must stay in Event Mode for another event iteration, starting a new
+    /// super-dense time instant.
     pub discrete_states_need_update: bool,
     /// The FMU requests to stop the simulation and the importer must call [`Common::terminate()`].
     pub terminate_simulation: bool,
-    /// At least one nominal value of the states has changed and can be inquired with [`ModelExchange::get_nominals_of_continuous_states()`].
-    /// This argument is only valid in Model Exchange.
+    /// At least one nominal value of the states has changed and can be inquired with
+    /// [`ModelExchange::get_nominals_of_continuous_states()`]. This argument is only valid in
+    /// Model Exchange.
     pub nominals_of_continuous_states_changed: bool,
     /// At least one continuous state has changed its value because it was re-initialized (see [https://fmi-standard.org/docs/3.0.1/#reinit]).
     pub values_of_continuous_states_changed: bool,
-    /// The absolute time of the next time event 𝑇next. The importer must compute up to `next_event_time` (or if needed
-    /// slightly further) and then enter Event Mode using [`Common::enter_event_mode()`]. The FMU must handle this time
-    /// event during the Event Mode that is entered by the first call to [`Common::enter_event_mode()`], at or after
+    /// The absolute time of the next time event 𝑇next. The importer must compute up to
+    /// `next_event_time` (or if needed slightly further) and then enter Event Mode using
+    /// [`Common::enter_event_mode()`]. The FMU must handle this time event during the Event
+    /// Mode that is entered by the first call to [`Common::enter_event_mode()`], at or after
     /// `next_event_time`.
     pub next_event_time: Option<f64>,
 }
