+++ src/fmi3/mod.rs
@@ -19,31 +19,35 @@ pub use fmi_schema::fmi3 as schema;
 pub enum Fmi3Res {
     /// The call was successful. The output argument values are defined.
     OK,
-    /// A non-critical problem was detected, but the computation may continue. The output argument values are defined.
-    /// Function logMessage should be called by the FMU with further information before returning this status,
-    /// respecting the current logging settings.
+    /// A non-critical problem was detected, but the computation may continue. The output argument
+    /// values are defined. Function logMessage should be called by the FMU with further
+    /// information before returning this status, respecting the current logging settings.
     Warning,
 }
 
 #[derive(Debug, thiserror::Error)]
 pub enum Fmi3Error {
-    /// The call was not successful and the FMU is in the same state as before the call. The output argument values are
-    /// undefined, but the computation may continue. Function logMessage should be called by the FMU with further
-    /// information before returning this status, respecting the current logging settings. Advanced importers may try
-    /// alternative approaches to continue the simulation by calling the function with different arguments or calling
-    /// another function - except in FMI for Scheduled Execution where repeating failed function calls is not allowed.
-    /// Otherwise the simulation algorithm must treat this return code like [`Fmi3Error::Error`] and must terminate the
-    /// simulation.
+    /// The call was not successful and the FMU is in the same state as before the call. The output
+    /// argument values are undefined, but the computation may continue. Function logMessage should
+    /// be called by the FMU with further information before returning this status, respecting the
+    /// current logging settings. Advanced importers may try alternative approaches to continue the
+    /// simulation by calling the function with different arguments or calling another function -
+    /// except in FMI for Scheduled Execution where repeating failed function calls is not allowed.
+    /// Otherwise the simulation algorithm must treat this return code like [`Fmi3Error::Error`]
+    /// and must terminate the simulation.
     ///
-    /// [Examples for usage of `Discard` are handling of min/max violation, or signal numerical problems during model evaluation forcing smaller step sizes.]
+    /// [Examples for usage of `Discard` are handling of min/max violation, or signal numerical
+    /// problems during model evaluation forcing smaller step sizes.]
     #[error("Discard")]
     Discard,
-    /// The call failed. The output argument values are undefined and the simulation must not be continued. Function
-    /// logMessage should be called by the FMU with further information before returning this status, respecting the
-    /// current logging settings. If a function returns [`Fmi3Error::Error`], it is possible to restore a previously
-    /// retrieved FMU state by calling [`set_fmu_state`]`. Otherwise [`FreeInstance`] or `Reset` must be called. When
-    /// detecting illegal arguments or a function call not allowed in the current state according to the respective
-    /// state machine, the FMU must return fmi3Error. Other instances of this FMU are not affected by the error.
+    /// The call failed. The output argument values are undefined and the simulation must not be
+    /// continued. Function logMessage should be called by the FMU with further information before
+    /// returning this status, respecting the current logging settings. If a function returns
+    /// [`Fmi3Error::Error`], it is possible to restore a previously retrieved FMU state by calling
+    /// [`set_fmu_state`]`. Otherwise [`FreeInstance`] or `Reset` must be called.  When detecting
+    /// illegal arguments or a function call not allowed in the current state according to the
+    /// respective state machine, the FMU must return fmi3Error. Other instances of this FMU are
+    /// not affected by the error.
     #[error("Error")]
     Error,
     #[error("Fatal")]
