+++ src/lib.rs
@@ -1,5 +1,5 @@
 //! The `fmi` crate implements a Rust interface to FMUs (Functional Mockup Units) that follow FMI
-//! Standard. This version of the library supports FMI2.0. See http://www.fmi-standard.org/
+//! Standard. This version of the library supports FMI 2.0 and 3.0. See http://www.fmi-standard.org/
 //!
 //! # Examples
 //!
@@ -26,41 +26,8 @@ pub enum Error {
     #[error("Error instantiating import")]
     Instantiation,
 
-    #[error("Invalid fmi2Status {}", status)]
-    InvalidStatus { status: u32 },
-
-    /// For ME: It is recommended to perform a smaller step size and evaluate the model equations
-    /// again, for example because an iterative solver in the model did not converge or because a
-    /// function is outside of its domain (for example sqrt(<negative number>)). If this is not
-    /// possible, the simulation has to be terminated.
-    ///
-    /// For CS: fmi2Discard is returned also if the slave is not able to return the required status
-    /// information. The master has to decide if the simulation run can be continued.
-    #[error("fmi2Discard")]
-    FmiStatusDiscard,
-
-    /// The FMU encountered an error.
-    /// The simulation cannot be continued with this FMU instance.
-    /// If one of the functions returns fmi2Error, it can be tried to restart the simulation from
-    /// a formerly stored FMU state by calling fmi2SetFMUstate. This can be done if the capability
-    /// flag canGetAndSetFMUstate is true and fmu2GetFMUstate was called before in non-erroneous
-    /// state. If not, the simulation cannot be continued and fmi2FreeInstance or fmi2Reset must
-    /// be called afterwards.
-    #[error("fmi2Error")]
-    FmiStatusError,
-
-    /// The model computations are irreparably corrupted for all FMU instances.
-    /// [For example, due to a run-time exception such as access violation or integer division by
-    /// zero during the execution of an fmi function].
-    /// It is not possible to call any other function for any of the FMU instances.
-    #[error("fmi2Fatal")]
-    FmiStatusFatal,
-
     #[error("Unknown variable: {}", name)]
     UnknownVariable { name: String },
-
-    #[error("unknown toolchain version: {}", version)]
-    UnknownToolchainVersion { version: String },
 
     #[error("Model type {0} not supported by this FMU")]
     UnsupportedFmuType(String),
@@ -71,23 +38,15 @@ pub enum Error {
     #[error("Unsupported FMI version: {0}")]
     UnsupportedFmiVersion(String),
 
-    /*
-    #[error(
-        "TypesPlatform of loaded API ({:?}) doesn't match expected ({:?})",
-        found,
-        fmi2::fmi2TypesPlatform
-    )]
-    TypesPlatformMismatch { found: Box<[u8]> },
-    */
-    #[error(
-        "FMI version of loaded API ({:?}) doesn't match expected ({:?})",
-        found,
-        expected
-    )]
-    FmiVersionMismatch {
-        found: Box<[u8]>,
-        expected: Box<[u8]>,
-    },
+    //TODO: Fix
+    //#[error(
+    //    "TypesPlatform of loaded API ({:?}) doesn't match expected ({:?})",
+    //    found,
+    //    fmi2::fmi2TypesPlatform
+    //)]
+    //TypesPlatformMismatch { found: Box<[u8]> },
+    #[error("FMI version of loaded API ({found}) doesn't match expected ({expected})")]
+    FmiVersionMismatch { found: String, expected: String },
 
     #[error(transparent)]
     Io(#[from] std::io::Error),
@@ -101,9 +60,6 @@ pub enum Error {
     #[error(transparent)]
     Utf8Error(#[from] std::str::Utf8Error),
 
-    //#[cfg(feature = "fmi3")]
-    //#[error(transparent)]
-    //Fmi3ModelError(#[from] fmi3::model::ModelError),
     #[error(transparent)]
     LibLoading {
         #[from]
@@ -112,31 +68,33 @@ pub enum Error {
 
     #[cfg(feature = "fmi2")]
     #[error(transparent)]
-    Fmi2Error(#[from] fmi2::Fmi2Err),
+    Fmi2Error(#[from] fmi2::Fmi2Error),
+
+    #[cfg(feature = "fmi3")]
+    #[error(transparent)]
+    Fmi3Error(#[from] fmi3::Fmi3Error),
 }
 
 /// Ok Status returned by wrapped FMI functions.
 //#[derive(Debug, PartialEq)]
-//pub enum FmiStatus {
+// pub enum FmiStatus {
 //    Ok,
 //    Warning,
 //    Pending,
 //}
 
-/*
-impl From<fmi2::fmi2Status> for std::result::Result<FmiStatus, FmiError> {
-    fn from(fmi_status: fmi2::fmi2Status) -> Self {
-        match fmi_status {
-            fmi2::fmi2Status::OK => Ok(FmiStatus::Ok),
-            fmi2::fmi2Status::Warning => Ok(FmiStatus::Warning),
-            fmi2::fmi2Status::Discard => Err(FmiError::FmiStatusDiscard),
-            fmi2::fmi2Status::Error => Err(FmiError::FmiStatusError),
-            fmi2::fmi2Status::Fatal => Err(FmiError::FmiStatusFatal),
-            fmi2::fmi2Status::Pending => Ok(FmiStatus::Pending),
-        }
-    }
-}
-*/
+// impl From<fmi2::fmi2Status> for std::result::Result<FmiStatus, FmiError> {
+// fn from(fmi_status: fmi2::fmi2Status) -> Self {
+// match fmi_status {
+// fmi2::fmi2Status::OK => Ok(FmiStatus::Ok),
+// fmi2::fmi2Status::Warning => Ok(FmiStatus::Warning),
+// fmi2::fmi2Status::Discard => Err(FmiError::FmiStatusDiscard),
+// fmi2::fmi2Status::Error => Err(FmiError::FmiStatusError),
+// fmi2::fmi2Status::Fatal => Err(FmiError::FmiStatusFatal),
+// fmi2::fmi2Status::Pending => Ok(FmiStatus::Pending),
+// }
+// }
+// }
 
 pub mod built_info {
     // The file has been placed there by the build script.
